+++ README.md
@@ -2,7 +2,8 @@
 Challenges for LLM
 
 # Building
-Copy `conf_templates` to `conf` and fill out the settings you want.
+Copy `conf_templates` to `conf` and fill out the settings you want. These are the admin app's setting and aren't loaded by the regular app. The regular app has it's own settings file 
+at `setting.json` that sets up the redis and database connection. This is used by the admin app, but the admin app will attempt to use the username and password in it's setting file.
 
 In one terminal run `pip install -r requirements.txt`, then `uvicorn users:app --reload`
