+++ app/admin/auth.py
@@ -10,9 +10,9 @@ from fastapi.openapi.models import OAuthFlows as OAuthFlowsModel
 from jose import JWTError, jwt
 from passlib.context import CryptContext
 from pydantic import BaseModel
-from ..schema import User
+from ..public.schema import User
 from fastapi.templating import Jinja2Templates
-from .settings import settings
+from .settings import settings, SessionLocal
 
 templates = Jinja2Templates(directory="templates")
 auth_router = APIRouter()
